:green_heart: chore: Move seeding codes to db directory
