+++ backend/db/seeding/factory/frieindship.factory.ts
@@ -2,7 +2,7 @@
 
 import { faker } from '@faker-js/faker';
 
-import { User } from '../../src/entity/user.entity';
+import { User } from '../../../src/entity/user.entity';
 
 export default (user1: User, user2: User, accept: boolean, lastMessageTime?: Date) => {
   return {
