+++ backend/db/seeding/factory/game-history.factory.ts
@@ -1,7 +1,7 @@
 // file name should be <entityName>.factory.ts
 import { faker } from '@faker-js/faker';
 
-import { User } from '../../src/entity/user.entity';
+import { User } from '../../../src/entity/user.entity';
 
 export default (user1: User, user2: User) => {
   return {
